+++ host/src/print.js
@@ -1,5 +1,19 @@
+import add from 'remote/add'
+
 import _ from 'lodash'
 
-export default function printMe() {
-  console.log(_.join(['Hello', 'webpack'], ' '));
+function component() {
+  const div = document.createElement('div')
+  const btn = document.createElement('button');
+
+  btn.innerHTML = _.join(['Hi', 'index'], ' ');
+  btn.onclick = function print() {
+    console.log(add(2,3))
+  }
+
+  div.appendChild(btn)
+
+  return div;
 }
+
+document.body.appendChild(component());
