+++ host/webpack.config.js
@@ -1,37 +1,40 @@
-const path = require('path')
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+const HtmlWebPackPlugin = require("html-webpack-plugin");
 const ModuleFederationPlugin = require("webpack/lib/container/ModuleFederationPlugin");
 
 const remoteURL = 'http://localhost:8081'
 
 module.exports = {
   entry: {
-    index: './src/index.js',
-    print: './src/print.js',
+    index: './src/index.js'
   },
+
   output: {
     filename: '[name].bundle.js',
-    path: path.resolve(__dirname, 'dist'),
-    clean: true
+    clean: true,
+    publicPath: 'http://localhost:8082/'
   },
+
   plugins: [
     new ModuleFederationPlugin({
-      name: "host",
-      filename: "remoteEntry.js",
-      remotes: {
-        remote: `remote@${remoteURL}/remoteEntry.js`
-      }
+        name: "host",
+        filename: "remoteEntry.js",
+        remotes: {
+            remote: `remote@${remoteUrl}/remoteEntry.js`,
+        },
     }),
-    new HtmlWebpackPlugin({
-      title: 'Host'
-    })
+    new HtmlWebPackPlugin({
+        template: "./src/index.html",
+    }),
   ],
+
   devServer: {
-    port: 8082
+    port: 8082,
+    historyApiFallback: true,
   },
+
   devtool: 'inline-source-map',
+
   optimization: {
-    runtimeChunk: 'single',
     splitChunks: {
       chunks: 'all'
     }
